Moving schema inside src for both form and vanilla form. Creating sanity.json for form
