+++ packages/form/src/index.js
@@ -3,6 +3,6 @@ export {getFormField} from './form';
 export {default as RegexLookup} from './regex-lookup';
 export {default as validation} from './validation';
 
-export {default as formObject} from '../schemas/form';
-export {default as formField} from '../schemas/formfield';
-export {default as formArrayField} from '../schemas/formarrayfield';
+export {default as formObject} from './schemas/form';
+export {default as formField} from './schemas/formfield';
+export {default as formArrayField} from './schemas/formarrayfield';
