+++ packages/vanilla-form/src/index.js
@@ -3,6 +3,6 @@ export {getVanillaFormField} from './vanilla-form';
 export {default as RegexLookup} from './regex-lookup';
 export {default as validation} from './validation';
 
-export {default as formObject} from '../../form/schemas/form';
-export {default as formField} from '../../form/schemas/formfield';
-export {default as formArrayField} from '../../form/schemas/formarrayfield';
+export {default as formObject} from './schemas/form';
+export {default as formField} from './schemas/formfield';
+export {default as formArrayField} from './schemas/formarrayfield';
